feat(local-common): Add shared queue message dispatch handler
Move processQueueMessage from @ez4/local-queue into @ez4/local-common
so both in-memory and ElasticMQ emulators can use the same dispatch
lifecycle. Rename from processLambdaMessage to processQueueMessage
and keep local-queue compatibility via re-export.
