+++ providers/local/local-common/src/main.ts
@@ -1,4 +1,5 @@
 export * from './emulators/module';
 
+export * from './handlers/queue';
 export * from './utils/dispatch';
 export * from './utils/response';
